+++ base/deprecated.jl
@@ -1599,6 +1599,12 @@ end
 # issue #6466
 # `write` on non-isbits arrays is deprecated in io.jl.
 
+# issue #22791
+@deprecate_binding select partialsort
+@deprecate_binding select! partialsort!
+@deprecate_binding selectperm partialsortperm
+@deprecate_binding selectperm! partialsortperm!
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations